+++ sim/state.go
@@ -102,7 +102,6 @@ type CommonState struct {
 	Facility          string
 	MagneticVariation float32
 	NmPerLongitude    float32
-	PrimaryAirport    string
 
 	SimDescription string
 	ScenarioBrief  string
@@ -302,7 +301,6 @@ func newCommonState(config NewSimConfiguration, startTime time.Time, model *wx.M
 		Facility:          config.Facility,
 		MagneticVariation: config.MagneticVariation,
 		NmPerLongitude:    config.NmPerLongitude,
-		PrimaryAirport:    config.PrimaryAirport,
 		SimDescription:    config.Description,
 
 		TFRs: config.TFRs,